Change name of Graphql Pojo classes
